+++ projects/angular2-smart-table/src/lib/components/tbody/cells/custom.component.ts
@@ -4,25 +4,31 @@ import {Row} from '../../../lib/data-set/row';
 import {Grid} from '../../../lib/grid';
 import {CustomAction} from '../../../lib/settings';
 import {CustomActionEvent} from '../../../lib/events';
+import {SecurityTrustType} from "../../../pipes/bypass-security-trust.pipe";
 
 @Component({
   selector: 'angular2-st-tbody-custom',
   changeDetection: ChangeDetectionStrategy.OnPush,
   template: `
-    <ng-container *ngFor="let action of customActions">
-      <a href="#" class="angular2-smart-action angular2-smart-action-custom-custom"
-         *ngIf="!action.renderComponent"
-         [innerHTML]="action.title"
-         (click)="onCustom(action, $event)"></a>
-      <a href="#" class="angular2-smart-action angular2-smart-action-custom-custom"
-         *ngIf="action.renderComponent"
-         (click)="onCustom(action, $event)">
-        <angular2-st-tbody-custom-item class="angular2-smart-action angular2-smart-action-custom-custom"
-                                  [action]="action"
-                                  [row]="row"></angular2-st-tbody-custom-item>
-      </a>
-
-    </ng-container>
+      <ng-container *ngFor="let action of customActions">
+          <a href="#" class="angular2-smart-action angular2-smart-action-custom-custom"
+             *ngIf="!action.renderComponent && showAction(action)"
+             [ngClass]="{'not-allowed': disableAction(action)}"
+             [innerHTML]="buttonContent(action) | bypassSecurityTrust: bypassSecurityTrustFor(action)"
+             (click)="onCustom(action, $event)"
+          ></a>
+          <a href="#" class="angular2-smart-action angular2-smart-action-custom-custom"
+             *ngIf="action.renderComponent && showAction(action)"
+             [ngClass]="{'not-allowed': disableAction(action)}"
+             (click)="onCustom(action, $event)"
+          >
+              <angular2-st-tbody-custom-item
+                      class="angular2-smart-action angular2-smart-action-custom-custom"
+                      [action]="action"
+                      [row]="row"
+              ></angular2-st-tbody-custom-item>
+          </a>
+      </ng-container>
   `,
 })
 export class TbodyCustomComponent {
@@ -36,6 +42,22 @@ export class TbodyCustomComponent {
     return this.grid.getSetting('actions.custom') ?? [];
   }
 
+  buttonContent(action: CustomAction): string {
+    return action.customButtonContent ?? action.title ?? action.name;
+  }
+
+  bypassSecurityTrustFor(action: CustomAction): SecurityTrustType {
+    return (action.sanitizer?.bypassHtml ?? false) ? 'html' : 'none';
+  }
+
+  showAction(action: CustomAction): boolean {
+    return action.hiddenWhen === undefined || !action.hiddenWhen(this.row);
+  }
+
+  disableAction(action: CustomAction): boolean {
+    return action.disabledWhen !== undefined && action.disabledWhen(this.row);
+  }
+
   onCustom(action: CustomAction, event: MouseEvent) {
     event.preventDefault();
     event.stopPropagation();
@@ -47,5 +69,4 @@ export class TbodyCustomComponent {
       source: this.source,
     });
   }
-
 }